src/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
strands-agents
strands-agents-tools
mcp
boto3
python-dotenv
requests
watchtower
pytest
pytest-asyncio
